+++ packages/terra-application-name/src/terra-dev-site/test/application-name/ApplicationMenuNameDefault.test.jsx
@@ -3,7 +3,7 @@ import Image from 'terra-image';
 import ApplicationMenuName from '../../../ApplicationMenuName';
 
 export default () => (
-  <div style={{ backgroundColor: 'green', width: '100%' }}>
+  <div style={{ width: '100%' }}>
     <ApplicationMenuName
       id="default"
       title="Title"