+++ packages/terra-application-utility/tests/wdio/application-menu-utility-spec.js
@@ -11,23 +11,24 @@ describe('ApplicationMenuUtility', () => {
     Terra.should.beAccessible({ viewports, context: '#default' });
     Terra.should.themeCombinationOfCustomProperties({
       testName: 'themed',
+      selector: '#default',
       properties: {
-        '--terra-application-menu-utility-background-color': 'blue',
-        '--terra-application-menu-utility-border-bottom-width': '6px',
-        '--terra-application-menu-utility-border-left-width': '6px',
-        '--terra-application-menu-utility-border-radius': '6px',
-        '--terra-application-menu-utility-border-right-width': '6px',
-        '--terra-application-menu-utility-border-top': '6px solid blue',
-        '--terra-application-menu-utility-color': 'purple',
-        '--terra-application-menu-utility-padding-left': '20px',
-        '--terra-application-menu-utility-padding-right': '20px',
-        '--terra-application-menu-utility-icon-min-width': '15px',
-        '--terra-application-menu-utility-accessory-height': '2rem',
-        '--terra-application-menu-utility-accessory-margin-right': '20px',
-        '--terra-application-menu-utility-accessory-width': '2rem',
-        '--terra-application-menu-utility-title-font-size': '2rem',
-        '--terra-application-menu-utility-title-font-weight': 'bold',
-        '--terra-application-menu-utility-title-margin-right': '20px',
+        '--terra-application-utility-menu-background-color': 'blue',
+        '--terra-application-utility-menu-border-bottom-width': '6px',
+        '--terra-application-utility-menu-border-left-width': '6px',
+        '--terra-application-utility-menu-border-radius': '6px',
+        '--terra-application-utility-menu-border-right-width': '6px',
+        '--terra-application-utility-menu-border-top': '6px solid blue',
+        '--terra-application-utility-menu-color': 'purple',
+        '--terra-application-utility-menu-padding-left': '20px',
+        '--terra-application-utility-menu-padding-right': '20px',
+        '--terra-application-utility-menu-icon-width': '15px',
+        '--terra-application-utility-menu-accessory-height': '2rem',
+        '--terra-application-utility-menu-accessory-margin-right': '20px',
+        '--terra-application-utility-menu-accessory-width': '2rem',
+        '--terra-application-utility-menu-title-font-size': '2rem',
+        '--terra-application-utility-menu-title-font-weight': 'bold',
+        '--terra-application-utility-menu-title-margin-right': '20px',
       },
     });
   });
@@ -37,7 +38,7 @@ describe('ApplicationMenuUtility', () => {
     Terra.should.matchScreenshot({ viewports, selector: '#default' });
     Terra.should.beAccessible({ viewports, context: '#default' });
     Terra.should.themeEachCustomProperty('#default', {
-      '--terra-application-menu-utility-hover-background-color': 'blue',
+      '--terra-application-utility-menu-hover-background-color': 'blue',
     });
   });
 
@@ -46,7 +47,7 @@ describe('ApplicationMenuUtility', () => {
     Terra.should.matchScreenshot({ viewports, selector: '#default' });
     Terra.should.beAccessible({ viewports, context: '#default' });
     Terra.should.themeEachCustomProperty('#default', {
-      '--terra-application-menu-utility-focus-box-shadow': '0 0 4px 4px rgba(80, 80, 233, 0.5), 0 0 6px 7px rgba(80, 80, 233, 0.35)',
+      '--terra-application-utility-menu-focus-box-shadow': '0 0 4px 4px rgba(80, 80, 233, 0.5), 0 0 6px 7px rgba(80, 80, 233, 0.35)',
     });
   });
 });