+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogOnModalManager.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import ModalManager from 'terra-modal-manager';
+import NotificationDialogModalContainer from '../../doc/common/NotificationDialogModalContainer';
+
+const NotificationDialogOnModalManager = () => (
+  <ModalManager>
+    <p> Notification Dialog has the highest z-index of 9001. Click the button to trigger Notification Dialog </p>
+    <NotificationDialogModalContainer />
+  </ModalManager>
+);
+
+export default NotificationDialogOnModalManager;