+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Added discernible text to footer Terra UI link
+
+2.1.0 - (February 5, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+2.0.0 - (January 25, 2019)
+------------------
+### Breaking Change
+* Updated variable names to match new naming standard
+* Updated terra-core dependencies
 
 1.16.0 - (January 5, 2019)
 ------------------