+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,19 @@ Changelog
 Unreleased
 ----------
 
+3.1.0 - (February 5, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+3.0.0 - (January 25, 2019)
+------------------
+### Breaking Change
+* Updated variable names to match new naming standard
+* Added new variables for css theming
+* Removed variables for css theming
+* Updated terra-core dependencies
+
 2.36.0 - (January 5, 2019)
 ------------------
 ### Changed